requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg[binary]==3.1.14

# Utilities
python-dotenv==1.0.0
orjson==3.9.10
xxhash==3.4.1
numpy==1.26.2
python-dateutil==2.8.2
httpx==0.25.2

# HTTP client
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
ruff==0.1.6
